Updated openstack/openstack Project: openstack/cinder  b608bd76c9fa70d7d355df346d4495dccac011b6
Dell SC: Incorrect values in REST API Login call

The Dell SC REST API driver (dell_storagecenter_api.py) could
specify the wrong api version header. On connection failure
the driver needs to parse the failure text to see if it is
the wrong value and try a second time.

Added the apiversion variable to the StorageCenterApiHelper
class so that this value only needs to be discovered once.

Updated open_connection appropriately.

Change-Id: Ibab6cf227f298b0fbf990aabcc674350a56c5439
Closes-Bug: 1548496