Updated openstack/openstack Project: openstack/neutron  d94ad2db0aeba855fb709a51d017e2eb13611f90
Add deprecation tag

We need to become more diligent and rigorous about the options and/or
system behaviors that we deprecate. We need to watch for field's
feedback and ensure that we provide an alternative route once a
deprecation is followed up by a removal.

Right now, there's no way to have a comprehensive view of what has
been deprecated in a cycle and what removals have to happen. Adding
this tag is the first step to get this sorted.

Change-Id: Id25ed35c8a23c0e6f6fc1b47545c8d1bab3e8dc7