Updated openstack/openstack Project: openstack-infra/project-config  2891d35e8e163f1510b644072c9e6ec8aadee606
Add neutron tags to some jobs

These will be passed through to the job, and can be stored by
subunit2sql and then used as a selector in the dashboard.

This set of jobs roughly corresponds to the jobs in the
short-lived periodic-neutron pipeline (though is slightly more
expansive in ovs).

Change-Id: Icf6e1b6b1c178330ac27aea51dddde666f98b2c2