Updated openstack/openstack Project: openstack/diskimage-builder  6d1bc6f1c38508297d8b40cac0be8b1d8e7b7d81
Refactor growroot for debuggabilty

This is a slight refactor I found useful when debugging.  The
udevadmin info query will be visbile in the output of "set -x" which
helps, and is the logs/journal.

We can also reduce some calls by keeping the value and just grepping
with a herefile.

This also does some error checking and bails out if it does not see
what it needs to continue.

Change-Id: I39c4d262f9c5ce53f6b83d95b1363a74834cf2c8