Updated openstack/openstack Project: openstack-infra/project-config  93bba96a60e85a1da43071879027d35a140cfa40
Add XStatic version checking job

For testing the proposed version checking job
it is being added to the experimental queue.

This check will be run in the release job as a pre-check
to ensure that the tag triggering the release matches
the version encoded in the release's files. If the version
coded in the tag doesn't match the version in the xstatic
package, it must be rejected.

Co-Authored-By: Joshua Hesketh <josh@nitrotech.org>
Change-Id: I04765ccc8c3a72758391ab147e7d5d79b828fd21