Updated openstack/openstack Project: openstack/api-site  90d57ddb15d967e49c84495f1506a1e72839ad84
[Orchestration API v1] Fix create_stack response params

Contrary to the docs, but consistent with the sample, the response to a
stack_create does not contain all of the data from the input. It looks as
if the documentation from the request was accidentally duplicated into the
response, and the documentation for the environment and disable_rollback
options were added into the wrong section.

Also rewrote the documentation of the "rel" member of a link so that it is
less incomprehensible/wrong, and moved it to indicate that it is not a
top-level key in the output.

Change-Id: I6b940feea82ce5efaa934a1096f7f7fd46208387