Updated openstack/openstack Project: openstack/neutron  23ea124f7930398f3e13d5f346e3ad06edbf6f09
Extend dicts with original model in create/update

This adjusts the logic in ML2 create port and create
network to use the original DB object created rather
than attempting to retrieve it from the DB again.
This eliminates the race window where the object could
be deleted out from underneath the create operations
resulting in a 404.

To achieve this it adds a create_port_db and
create_network_db method to db_base_plugin_v2 that
returns the original model instead of returning a
dictionary.

This also adds a skip to a unit test that was using DB
events to trigger a rollback and retry. This test was
causing failures with a FlushError conflict between an
SQL alchemy new instance and a persistent instance in
the session. This bug is fixed in the follow-up patch.

Closes-Bug: #1556139
Related-Bug: #1556178
Change-Id: I3c4b4cb4f63236fca24abaf58881879c2d5657fa