Updated openstack/openstack Project: openstack-infra/project-config  edf1308d8b11772468b58885e126737bc514daf5
Add centos-minimal build, experimental job

Add a build of centos-minimal to nodepool.  Let's just add this to
Rackspace first, because that's the "hard" place to get working
(configdrive, no dhcp, etc).  FWIW I have tested this image with my
RAX account.

The plan is to replace devstack-centos7 with this image globally once
we've got it stable.

Add an experimental job to use this new node for testing (devstack
should be able to bootstrap itself on these types of nodes).

Change-Id: I02fd98f5df7f90aa5d50666166b5c736f6f06aaf
Depends-On: I2956aaa49ba3137a799f97e0983ab4b7c93a0a0c