Updated openstack/openstack Project: openstack/governance  88a3e7bca3494371abf399f2884ec52c3e9865cb
Horizon plug-ins should not have type:library

Horizon plug-ins, while technically delivered as a Python
"library", are not considered a library as far as release
management deadlines are concerned. A few of them still
had the type:library tag, though, so let's fix that.

NB: we may want to introduce a type:horizon-plugin tag at
some point to designate those.

Change-Id: Idedf14a488372d7c2d4120f0637af7b05938b0b8