Updated openstack/openstack Project: openstack/heat  eaeda9892092995a43f87d0707a78f32e8eb21d4
Unit tests: compare stack names in PhysName

Previously we were ignoring the stack name when comparing a PhysName, since
that part can be shortened. That allowed a lot of code to get changed that
made the stack name out of sync with the one printed by the PhysName
object. This brings everything back into sync and starts comparing for
equality against the shortened stack name.

This also introduces a semi-correct implementation of __hash__.

Change-Id: I184cc7b0b746b1b7cddff231a90b7337ca42d35b