Updated openstack/openstack Project: openstack/cinder  7d6da7cda9e250faed7b682ea4e6bc92d8b9e892
Add the key 'replication' and set the correct 'replication_status'

This patch fixes the minor issues following the implementation
of replication v2.1 for storwize driver. It adds the key
'replication' to make sure of that the cinder-volume service
can have a successful restart and sets the 'replication_status
'back to enabled for failback.

Change-Id: I1ddb715e72336378434b7ee5ffe617cbb94cd16b
Closes-Bug: #1556285