Updated openstack/openstack Project: openstack/nova  5ae39f70f5b202bffd3165ecb9082d4a4f25090b
libvirt: Comment non-obvious security implications of migrate code

There are 2 lines in finish_migration which look like a bug. The
original reason for adding them no longer exists, but removing them
now would open a severe security bug due to a series of non-obvious
interactions. This patch simply adds a comment to ensure this is taken
into consideration when changing this code in future.

Change-Id: Icfe4b5deadfe7dbb0f34ea9af1f95b62cfca8a0a