Updated openstack/openstack Project: openstack/cinder  e0c98f0f688be35fb8da2ceda92541927472b422
Host selection in backup service

With scaling backup service, it introduces a config called
backup_use_same_backend which indicates whether customers use
same backup driver in their environments. If the value is
set to True, Cinder selects any backup host to do backup task. If
the value is False, cinder can only select backup.host for a backup
task. Currently the default value is set to False.

This patch is to change above config to backup_use_same_host which
means whether selecting same host for tasks of a backup. The
default value is set to False. It indicates that we don't care
about backup.host when running backup task, and choose any available
host in same az.

As currently Cinder doesn't support multiple backup drivers, the change
is to make the config more sensible.

Change-Id: I1a43df251a18006363162ee8e7c5aa891f44fa3a
Closes-bug: #1554845