Updated openstack/openstack Project: openstack/manila  c1f85e7a59885c7231223ffd96201dfc81f55498
Fix handling of share server details after error

Manila share manager has possibility to save share servers
details even if it failed to be created on some step.
To use it, share driver should attach 'detail_data' attr
to exception with dict of share server details.
But, if such data not provided or has unexpected type, then
manager writes "warnings" to LOG. So, make it have "debug" level
of logging because here is nothing to do for admin/operator.

Also, fix Generic driver's case when it does not provide information
when has it - on negative result of
share server network availability check.

Change-Id: If8926ad383669fd5ccbe69cf13b574bc044f19be
Closes-Bug: #1547076