Updated openstack/openstack Project: openstack/neutron-lbaas  a8c9ff6a5385c3967b6c09d5e6f9dcbbb022062f
Fix delete of session_persistence with pool update

Our LBaaSv2 database code was set to delete a pool's session_persistence
if the pool was updated and the update API request did not include
mention of the session_persistence. Expected behavior would be to not
change the session_persistence.

In addition, while editing this code it became obvious that we didn't
provide a good way to clear the session_persistence for a pool once it
was set, especially via the CLI. This has also been corrected.

Change-Id: I654b172927e1d96677a7da9e0846231b0ac48aa9
Depends-On: Idcf12e463fbaa3a61a211f13986d8472f52036d2
Closes-Bug: #1552086
Partial-Bug: #1547157