Updated openstack/openstack Project: openstack/api-site  0e71e8699b974b432290a11dde356d841eab743d
Fixes missing information in the Response Parameter table.

When listing all extensions using Shared File Systems API v2, The response
body contains the information of "links" field. However, in the Response
Parameter table, the information if "links" parameter is missed.
This patch add the information of "links" parameters to the Response
Parameters table

Change-Id: I2cd538660c4bad1e0dcf8d901c72ef173ab0f56f
Closes-Bug: #1553027