Updated openstack/openstack Project: openstack/api-site  85a3fd28a6ca53a3be6683351c7f2263303d5fb2
Missing information in the Response Parameter table in shared-api.

When creating a share using Shared File Systems API v2, the response body
contains the information of "links" field. However, in the Response
Parameters table, the information of "links" parameter is missed.
This patch adds the information of "links" parameter to the Response
Parameters table.

Change-Id: I62d93d39c7b42a8127d2b53145549bf55e842491
Closes-Bug: #1553053