Updated openstack/openstack Project: openstack/trove-integration  e702dda2b5295e1e96fea39605d9a5e427ac4e97
Fix the issue with DB2 install

DB2 installation fails because the hostname cannot be
correctly resolved. This was caused by the change in:
 https://review.openstack.org/#/c/272711/.

The change includes adding '127.0.0.1 <hostname>' to the
/etc/hosts file before the actual installation and deleting
this entry from /etc/hosts once the installation is complete.

To make this complete, a change is added to the guest
agent to update the db2 settings for hostname:
https://review.openstack.org/#/c/291820/

Change-Id: Ia9dad59b9b40e7df66f3b1659e529eb32bf91f17
Closes-Bug: #1556159