Updated openstack/openstack Project: openstack/manila  97c700430670b5de0583e46fddfe8ac1b0a2f268
HNAS: Enable no_root_squash option when allowing access to a share

Currently, the default option in HNAS driver is set root_squash on
when adding a rule to an export.
It prevents migration for working in Manila HNAS driver and make
impossible to change the ownership of files.
This patch enable the 'no_root_squash' option when allowing access
to the share.

Closes-bug: #1554148
Change-Id: Iaf40bed57590ca27bd1228da28621f68d0ef2d38