Updated openstack/openstack Project: openstack-infra/shade  2d9a7f054e0688c2c9c28638a9949c293eb4d0d0
Test v3 params on v2.0 endpoint; Add v3 unit

Part of the code was not being excericed resulting an invalid line of
code slipping through. "'%url' % interface" evaluates %u thinking it
is about to put in a number. This is not the case. Use .format()
instead.

Add unit test to check that section of code.

Additionally, add tests to validate v3 endpoint code

Change-Id: I116fff73ce102232c2a14acc7ee55bd660dd8be7