Updated openstack/openstack Project: openstack-infra/shade  98e63158e03844848ca24e1454bc636c5999ea2c
Update create_endpoint()

Using kwargs as it is used in create_endpoint() does not work if other
kwargs are needed besides the three *_url ones. This means if any new
kwargs come along this will have to be refactored anyway. More
immediately, this code cannot be used with update_endpoint() in the next
patchset leading to an inconsistency in code.

This refactor allows both create_endpoint and update_endpoint to be
consitent with eachother and, dare I say, makes this.... futureproof?

Change-Id: I79273e3dc409cf49b03cc644ea50fe143bcd53aa