Updated openstack/openstack Project: openstack/horizon  41533820a0330bc754ffd004c46b6578f8e4e7bb
Allow delete servers in error state

In instance list, when we have one/multiple servers in error state,
Horizon does not display "Terminate instances" button in top-right
side (close to "Launch instance" button). So, Horizon does not allow
terminate instances in error state.

When you click on "DELETE INSTANCE" button in Horizon, and something
went wrong (for example, the message to delete did not arrive to
Nova or Nova was stuck during the deletion process), you can see the
instance in "deleting" status forever. Right now, in this status,
the "delete instance" button disapears, so the user can not "retry"
the deletion.

Change-Id: I6486d6989d4bd89d5c9346125aa01ec1359352d4
Closes-bug: #1557353