Updated openstack/openstack Project: openstack-infra/devstack-gate  589ca4f1c28990fd51dc78382542b0388ff2ebbc
Fix bug in populating multiline env vars on reproduce

The script to populate reproduce.sh is failing when populating
multi-line env vars. That is the case of the enable_plugin lines,
when more than one plugin is enabled.
Add this workaround to properly populate all needed settings.

Also add a test to ensure that reproduce method is working
as expected.

Change-Id: I5287098a47162321efadbde7d069fa2a4e7eabf7