Updated openstack/openstack Project: openstack/oslo.log  c47a91dbbb586c27d8521b1016bf7901c47b1c90
Explicitly exclude tests from bandit scan

Now that the bandit linter runs without a bandit.yaml specified,
it should exclude the tests directory similar to how it did before
in the bandit.yaml.

The reason this has not caused a problem yet is because bandit is
still finding a config bandit.yaml in .tox/pep8/etc/bandit/bandit.yaml.
But in a newer version of bandit this will not be the case.

Change-Id: I4fd5b0cc60309dfc207cdf2025caae37836391da