Updated openstack/openstack Project: openstack/oslo.messaging  41851e9c8a2eb9c16dee899c28b99639d89e5ec4
Explicitly exclude tests from bandit scan

Now that the bandit linter runs without a bandit.yaml specified,
it should exclude the tests directory similar to how it did before
in the bandit.yaml.

The reason this has not caused a problem yet is because bandit is
still finding a config bandit.yaml in .tox/pep8/etc/bandit/bandit.yaml.
But in a newer version of bandit this will not be the case.

Change-Id: I6cdceaff2aa906e813f1d1e993f60e66358946ee