Updated openstack/openstack Project: openstack/horizon  0c5c57542b903debbecf684db88e84a7a6fe863d
Add support for detail actions

This updates the angular actions directive to support the "detail"
list type to render actions as tiles with a title and description.

To see it in action:
Update some row actions so the template attribute has a title and
description property. Then add these actions to an array in the
detail page controller and add the actions directive to the content
of the detail page, something like this:

  <actions allowed="ctrl.detailActions" type="detail"
           item="ctrl.item" ng-if="ctrl.item"></actions>

Implements blueprint next-steps
Change-Id: I5e85004255e351c1fdd121251030e0697b3b9b9f