Updated openstack/openstack Project: openstack-infra/tripleo-ci  d2757c5739bd763a8e2b46f8a0ca27e44b041ad8
Move tripleo.sh into tripleo-ci repo

We're discovering that everything needs backporting into both branches
despite the script handling both branches anyway.  Given that this
is designed only for developer and CI usage, we may as well save this
effort and move the script here instead.

When this is all working, we can remove the script from tripleo-common
and add a README instead pointing developers to the new location.

This is based on the content of tripleo-common at 6031198 - tripleo.sh
is unchanged except for adjusting the TENANT_PINGTEST_TEMPLATE path.

Change-Id: I9f3df108514a9d5f456e1c467993b778a2f91d89