Updated openstack/openstack Project: openstack/ironic-python-agent  c9f7c951dcd223caeadd27e7ddbba7a5169e9f2f
Stop trying to log stdout when fetching logs during inspection

Logging the whole journalctl output is not the best idea. Fortunately,
it does not work right now and fails with a traceback :)

This change adds a new log_stdout argument to utils.execute() and uses it in
the "logs" inspection collector.

Also do not log the logs while logging the collected data.

Change-Id: Ibc726ac2c4f5eb06c73ac4765bb400077b84a6cc