Updated openstack/openstack Project: openstack/cinder  ea9fe360fbbed8a1f3c17439e11a40cd211a2233
NetApp E-Series: Volumes not added to consisgroup

Volumes were not being added to a consistency group upon creation when the
consistencygroup_id is provided during the volume creation.

This patch adds functionality to relevant methods in library.py to add the newly
created volume to a consistency group when applicable.

Closes-Bug: #1555933
Change-Id: I6d09b81571ce7b5c07d155cb4a70079fe35ed945