Updated openstack/openstack Project: openstack/cinder  42c541534ef016256070e5501a5e07c7469fdeeb
Huawei: Creating hypermetro failed in the remote pool

Create the hypermetro failed when use the remote pools on the
huawei driver, the log print get pool id fail. The 'pool_id'
is not exist in the LUN property dictionary, so the 'pool_id'
should be modified to 'PARENTID' to save the remote pool id.

Closes-Bug:#1547798
Change-Id:Ic99531dad4fab5e22e295e19bbfa94c789247d93