Updated openstack/openstack Project: openstack-infra/elastic-recheck  299bb40836d81c9afb3dab86f130f680276ea127
Add a cluster health indicator to the graph output

This uses the ES health API to get the cluster health status
and pretty it up in our index/gate graph pages.

The cluster health API is documented here:

https://www.elastic.co/guide/en/elasticsearch/reference/current/cluster-health.html

Also add a note to the readme on how to view the openstack
ES cluster health.

Change-Id: I3df833cf5024af7282e2602e7dfa5db9c3384b6a