Updated openstack/openstack Project: openstack/nova  bc5035343d366a18cae587f92ecb4e871aba974a
Add a cell and host mapping utility to nova-manage

This enables an admin to create a cell mapping and corresponding
host mappings in the API database. For example:

  nova-manage --config-file nova-cell.conf cell_v2 map_cell_and_hosts

With the following options:

  --transport-url (required) - The transport url for the cell MQ
  --name (optional) - The name of the cell
  --verbose (optional) - Output the uuid of the created cell

There is no release note since the command is only to be used
experimentally and for testing at this time.

Co-Authored-By: Sylvain Bauza <sbauza@redhat.com>

Change-Id: I7a29782e7cbafddd90a627992cac1f7f47045a36