Updated openstack/openstack Project: openstack/tempest  36521463de6b5c6c48f1e6d0ff987a074e2af406
Add note a section to lib doc about where to put plugins

This commit adds a subsection to the creating a plugin section in the
plugin doc that tries to strongly advocate creating a separate python
package for tempest plugins for various reasons. This will hopefully
steer potential plugin authors towards doing this.

Change-Id: Id3c06381c0d8e79f59e50486a3c10b12bb977c51