Updated openstack/openstack Project: openstack/swift  f2e344a4d98606385f34d3633a90c24aa32fc43d
Faster suffix invalidations on object PUT/DELETE

Before this commit, we performed a full read-modify-write cycle on a
partition's hashes.pkl to invalidate the relevant suffix whenever an
object PUT or DELETE completes.

Now we append invalid suffixes to a new file, "hashes.invalid",
alongside hashes.pkl. When we actually get a REPLICATE request and
need to compute the hashes, *then* we perform a full read-modify-write
on hashes.pkl and clear out hashes.invalid.

Change-Id: Ia7add438e25688b4b286f0110a4e43490e11ad75