Updated openstack/openstack Project: openstack/python-keystoneclient  8e9a810ebc3466ca7eb8505f74323b740de3d295
Enhance functional class to provide default info

Modify the base class for functional tests so that base scope and user
information can be loaded from the client, which in turn was
instantiated from cloud config.

The base scope and user info will be used in tests, e.g when creating a
user, we will use the domain from the base class to put the user in.
The base project will be used as the default user's project.

Change-Id: Ib01da8d8bfc11a06bf0d79fac2089ca5b9506670