Updated openstack/openstack Project: openstack/cinder  46fb4fcd4cb82aab2c66d93451970ce957d95daa
LeftHand: Add default SSH timeout and key values

Currently if ssh_conn_timeout or san_private_key is not specified
in the remote_array object, the client creation fails. This should not
be the case. Instead, we can add default values to allow the creation
to succeed.

In addition, this patch removes some redudent lines of code and
moves a call to a more logical place.

Change-Id: I0bb32c2ea9a16ea568059ff7a5a5f07a95c718eb
Closes-Bug: #1556331