Updated openstack/openstack Project: openstack/horizon  4d0ce35ded4112dbd63babc3cec73a26e705d280
Fix network duplication check logic

Previously api.neutron.Network object was compared by using _apidict
but it did not work when for example the order of subnet list is different.
This commit changes the comparison logic to use network ID explicitly.
In addition, it moves the logic to fetch external networks to api.neutron.

Change-Id: Ie3a42e29c32c17a7f3bf1596b0e09cb73eae9d2a
Closes-Bug: #1528465