Updated openstack/openstack Project: openstack/nova  0955914c6e793364310d35d2ee5efa4b8779ce56
Fix reno for RC1

This is a bus for all reno problems we see with the relnotes :

* Change Id25bd4870c6e2fda08dc0177b7ed61a8a6091838 was having
a formatting problem with not retaining the newlines, and also
was mentioning "Nove" :-)

* Change I43ddc5164f88fb049fe49ad4672fbf09b2b1987e was using a
prelude section that is not needed because we will provide a
general prelude section (and commenting that) in the next change.

* Change I2dba71e71b1ed7cf8476e8bfe9481e84be5df128 was incorrectly
using the upgrades section while it was a deprecation.

* Change Ia6f1d8f9c7d700498b6441a6ba37fa25b2611ef8 is no longer
valid because Id7936be290b6febd18deb4c2db8ea4d678d4d9b1 has been
merged during Mitaka too.

* Change I2465e1bee618a552132d659014ea96cdbda161d6 was reverting
a reno file but due to a reno bug (see
https://review.openstack.org/#/c/293078/), the release note is
still kept in the output. Consequently, the workaround I found
is to reintroduce the same reno file with an empty content to
drop the previous content.

Note that I used the prelude section for "dropping" notes, that
prevents us to provide empty bullets for upgrade or features sections.

Change-Id: I1535aff80850fa3666da739133fc43e8579aa19b