Updated openstack/openstack Project: openstack/manila  301584f3ab340cb4a207cdae5cd6f77e510a7597
service instance: also recognize instance name

Using the generic driver needs a service instance. The variable
"service_instance_name_or_id" indicates that it's possible to use
the UUID or the name of the Nova VM. Using the VM name currently only
works, when the instance is in the current tenant.
This fix searches the instance at first in the current tenant and
if not found, it searches again over all tenants.
Without this fix, you get the following error when using the name:

Failed to get Nova VM. No server with a name or ID of 'myname' exists.

Change-Id: I8770eff9752b76ebe490abf03a7ea6548282bf9b
Closes-Bug: #1535775