Updated openstack/openstack Project: openstack/oslo.middleware  1e5f746ac4d991d76505dd43b0da9594da5005c7
Set default allow methods to those defined in RFC 2616

This patch updates the default allow_methods config setting
to all headers defined in RFC 2616 Section 9. We were
cherry-picking headers here with no justification, and in order
to be backwards compatible, this patch opts to be inclusive
of all valid methods. Specific methods can still be overridden.

https://www.w3.org/Protocols/rfc2616/rfc2616-sec9.html

Change-Id: I3e87e3bdb643706bcd4def1780d84ac8d45addb9