Updated openstack/openstack Project: openstack/nova  99ad55ad9d0e954b8a175a61271a6f150cc05039
Fix reno reverts that are still shown

There is a reno bug (https://review.openstack.org/#/c/293078/) that leaves
reno files to be shown as release notes even if the file was deleted later.

This change aims to reintroduce a couple of reno files but emptied in order
to remove their previous content from the HTML output.
This is a workaround and those empty YAML files will be deleted once the
latest reno version is having the bugfix.

* Change I11bde778e9fe1f3a70d9fac213b40f05f07e7e47 was removing lock_policy
* Change Ic9f70dae037d32980a5a252bdd08eff02ba27120 was removing a filter
* Change I817b8d0f6c6fa71dc56b031c717bd7a63193f847 was removing report opts

There was one last revert, but it has already been fixed in
I1535aff80850fa3666da739133fc43e8579aa19b

Change-Id: I945ded1eceeac49f50a5fff7ebece040f8c8b632