Updated openstack/openstack Project: openstack/nova  a8a371bd6f5d640fc5ad8c6161571481d073c570
Add regression test for Cinder 403 forwarding

The Nova API should be returning a 403 whenever it gets a Forbidden from
Cinder. Nova currently only returns a 403 when it runs into a policy
exception from its own policy (which is a subclass of Forbidden). This
changes the handler over to return a 403 when any Forbidden comes in.

Also, the volume helper methods in the function API client had incorrect
URLs, so those were also fixed (/volumes --> /os-volumes)

Change-Id: If9dd002454338d07fcea0c2092c12f3961c3dcdf
Partial-Bug: #1554631
Closes-Bug: #1555826