Updated openstack/openstack Project: openstack/keystone  a3b38ee041e82268986b1ccccaad03da0a332703
Check for already present user without inserting in Bootstrap

keystone-manage bootstrap check for already present info in the DB
by trying an insert and catching a Conflict exception.
This will not work with databases where inserting an user is not
possible. Changed the code to try a get first, and insert when
the user is not found.

Change-Id: If15c284aae5d10c594688c588dde9b21675ff487
Closes-Bug: 1553216